REFACTOR(coreclient): signleton entrance
