+++ internal/stream/block/fetch_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/FissionAndFusion/lws/internal/coreclient"
 	"github.com/FissionAndFusion/lws/internal/coreclient/DBPMsg/go/dbp"
 	"github.com/FissionAndFusion/lws/internal/coreclient/DBPMsg/go/lws"
+	cclientModule "github.com/FissionAndFusion/lws/internal/coreclient/instance"
 	blockService "github.com/FissionAndFusion/lws/internal/db/service/block"
-	cclientModule "github.com/FissionAndFusion/lws/internal/stream/cclient"
 	"github.com/FissionAndFusion/lws/test/helper"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/any"
